refactor(team): regroup headless utilities by concern (C21)
Apply the "helpers next to their state" learning to headless_codex.go.
Three semantic moves shrink it from 297 to 178 lines and put each
helper next to the code that calls it.

* launchHeadlessCodex -> launcher_boot.go. It's a Launcher boot
  path, parallel to Launch (tmux mode) and LaunchWeb (web mode).
  Belongs with the other boot methods, not buried in the codex
  dispatch file.

* wuphfLogDir / wuphfLogDirOverride / appendHeadlessCodexLog /
  appendHeadlessCodexLatency -> new headless_logging.go (74 lines).
  These are the file-based logging utilities used by every headless
  turn runner (codex, claude, opencode, openai-compat). They have no
  codex-specific logic; the codex prefix in the function names
  reflects history, not current scope.

* durationMillis -> headless_progress.go. It computes the
  millisecond delta the headlessProgressMetrics struct stores; the
  two pieces are one conceptual unit.

* headlessCodexTaskID -> headless_codex_queue.go. The queue is the
  only caller — it parses the prompt for #task-/#blank-slate-
  prefixes to drive dedup decisions in enqueueHeadlessCodexTurnRecord
  and replaceDuplicateTaskTurnLocked.

After the moves, headless_codex.go owns exactly: the cross-provider
dispatch entry (defaultHeadlessCodexRunTurn / headlessCodexRunTurn),
the override seams, the timeout-env helper + timeout consts, the
shared queue types (headlessCodexTurn, headlessCodexActiveTurn,
headlessWorkerPool), and the workspace-status snapshot seam used by
the durability check. Codex-specific scaffolding only.

headless_codex.go: 297 -> 178 (-119). New: headless_logging.go (74),
durationMillis appended to headless_progress.go, headlessCodexTaskID
appended to headless_codex_queue.go, launchHeadlessCodex appended to
launcher_boot.go.

Stacked on refactor/launcher-c20-lifecycle-phases (C20).
